refactor(pages): reformat code and optimize performance
