+++ src/App.css
@@ -1,52 +1,45 @@
 @import "../node_modules/normalize.css/normalize.css";
 @import '~@fortawesome/fontawesome-free/css/all.min.css';
-@import url('https://fonts.googleapis.com/css2?family=Space+Grotesk:wght@300..700&display=swap');
 
 :root {
-  --primary-color:#0085ca;
-  --accent-color:#002f6c;
+    --primary-color:#0085ca;
+    --accent-color:#002f6c;
 }
 
-*{
-  margin: 0;
-  padding: 0;
-}
-
-h1{
-  margin: 0;
+* {
+    margin: 0;
+    padding: 0;
 }
 
 .App {
-  text-align: center;
-  font-family: "Space Grotesk", monospace;
+    text-align: center;
+    font-family: "Space Grotesk", monospace;
 }
 
-.primary-button{
-  font-size: 1rem;
-  padding: 1rem;
-  border-radius: 0.3rem;
-  background-color: #0085ca;
-  color: #ffffff;
-  border:none;
-  text-decoration: none;
-  transition: all 0.3s;
+.primary-button {
+    font-size: 1rem;
+    padding: 1rem;
+    border-radius: 0.3rem;
+    background-color: #0085ca;
+    color: #ffffff;
+    border:none;
+    text-decoration: none;
+    transition: all 0.3s;
+
+    &:hover {
+        background-color: #002f6c;
+        cursor: pointer;
+    }
 }
 
-.primary-button:hover{
-  background-color: #002f6c;
-  cursor: pointer;
-}
-
-.display-none{
-  display: none !important;
+.display-none {
+    display: none !important;
 }
 
 body {
-  overflow: scroll;
+    overflow: scroll;
+
+    &::-webkit-scrollbar {
+        display: none;
+    }
 }
-
-body::-webkit-scrollbar {
-  display: none;
-}
-
-
