+++ src/App.jsx
@@ -36,7 +36,7 @@ function App() {
 
     return (
         <div className="App">
-        <RouterProvider router={router} />
+            <RouterProvider router={router} />
         </div>
     );
 }
