+++ src/pages/ContactPage/ContactPage.css
@@ -1,5 +1,5 @@
-.contact-page-containers{
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+.contact-page-containers {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
 }
