+++ src/pages/CoursesPage/CoursesPage.css
@@ -1,4 +1,4 @@
-.courses-page-container{
+.courses-page-container {
     margin-top: 5rem;
     display: flex;
     flex-direction: column;
