+++ src/pages/EventsPage/EventsPage.css
@@ -1,4 +1,4 @@
-.events-page-container{
+.events-page-container {
     margin-top: 5rem;
     width: 100%;
     display: flex;
