+++ src/pages/HomePage/HomePage.css
@@ -1,4 +1,4 @@
-.home-page-container{
+.home-page-container {
     width: 100%;
     display: flex;
     flex-direction: column;
