+++ src/pages/HomePage/HomePage.jsx
@@ -8,20 +8,21 @@ import WhatsNext from "../../components/WhatsNext/WhatsNext";
 import { useLoaderData } from "react-router-dom";
 
 function HomePage() {
-  const homeData = useLoaderData();
+    const homeData = useLoaderData();
 
-  useEffect(() => {
-    window.scrollTo(0, 0);
-  }, []);
-  return (
-    <div className="home-page-container">
-      <HomeIntro />
-      <Stats />
-      <Sponsors />
-      <WhatsNext events={homeData} />
-      <AboutHomePage />
-    </div>
-  );
+    useEffect(() => {
+        window.scrollTo(0, 0);
+    }, []);
+
+    return (
+        <div className="home-page-container">
+            <HomeIntro />
+            <Stats />
+            <Sponsors />
+            <WhatsNext events={homeData} />
+            <AboutHomePage />
+        </div>
+    );
 }
 
 export default HomePage;
